+++ packages/shuttle/src/shuttle/eventStream.ts
@@ -236,6 +236,7 @@ export class HubEventStreamConsumer extends TypedEmitter<HubEventStreamConsumerE
   private log: pino.Logger;
   private beforeProcess?: PreProcessHandler;
   private afterProcess?: PostProcessHandler;
+  private interval?: NodeJS.Timeout;
 
   constructor(
     hub: HubClient,
@@ -263,25 +264,29 @@ export class HubEventStreamConsumer extends TypedEmitter<HubEventStreamConsumerE
     this.stopped = false;
     await this.stream.waitUntilReady();
     await this.stream.createGroup(this.streamKey, this.groupName);
+
+    this.interval = setInterval(() => {
+      // Run separately from the run loop since it consumes wall clock time and
+      // doesn't need to happen often, as long as it happens at some frequency
+      this.clearOldEvents();
+
+      // While we do process stale events regularly, it's possible that a stream
+      // will be so busy that we won't have had an opportunity, so make sure we
+      // guarantee at least one invocation on some cadence.
+      this.processStale(onEvent);
+    }, 1000 * 60 /* 1 min */);
+
     void this._runLoop(onEvent);
   }
 
   stop() {
+    clearInterval(this.interval);
     this.stopped = true;
   }
 
   private async _runLoop(onEvent: (event: HubEvent) => Promise<Result<ProcessResult, Error>>) {
     while (!this.stopped) {
       try {
-        const sizeStartTime = Date.now();
-        const size = await this.stream.streamSize(this.streamKey);
-        statsd.gauge("hub.event.stream.size", size, { hub: this.hub.host, source: this.shardKey });
-        const sizeTime = Date.now() - sizeStartTime;
-
-        statsd.timing("hub.event.stream.size_time", sizeTime, {
-          hub: this.hub.host,
-          source: this.shardKey,
-        });
         let eventsRead = 0;
 
         const startTime = Date.now();
@@ -384,9 +389,7 @@ export class HubEventStreamConsumer extends TypedEmitter<HubEventStreamConsumerE
 
         if (eventsRead === 0) {
           if (this.stopped) break;
-          const numProcessed = await this.processStale(onEvent);
-          const numCleared = await this.clearOldEvents();
-          if (numProcessed + numCleared === 0) await sleep(10); // No events, so wait a bit to prevent CPU thrash
+          await this.processStale(onEvent);
         }
       } catch (e: unknown) {
         this.log.error(e, "Error processing event, skipping");